Fixing issues with incorrect implementation of LDAPUserDetails
