+++ example/example.gradle
@@ -17,6 +17,7 @@ dependencies {
     implementation(
         project(':authenticator'),
         project(':localauth'),
+        project(':localldap'),
         project(':ldap'),
         project(':cameo'),
         project(':elastic'),
