+++ localauth/src/main/java/org/openmbee/mms/localauth/security/LocalUsersDetailsService.java
@@ -1,22 +1,46 @@
 package org.openmbee.mms.localauth.security;
 
 import java.util.List;
+import java.util.Optional;
 
+import org.openmbee.mms.core.exceptions.ForbiddenException;
 import org.openmbee.mms.data.domains.global.User;
-import org.openmbee.mms.users.security.DefaultUsersDetailsService;
-import org.openmbee.mms.users.security.UsersCreateRequest;
+import org.openmbee.mms.users.security.AbstractUsersDetailsService;
+import org.openmbee.mms.users.objects.UsersCreateRequest;
 import org.openmbee.mms.users.security.UsersDetailsService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-public class LocalUsersDetailsService extends DefaultUsersDetailsService implements UsersDetailsService {
+public class LocalUsersDetailsService extends AbstractUsersDetailsService implements UsersDetailsService {
 
 
+    protected PasswordEncoder passwordEncoder;
 
+    @Autowired
+    public void setPasswordEncoder(PasswordEncoder passwordEncoder) {
+        this.passwordEncoder = passwordEncoder;
+    }
 
-    public List<User> getUsers() {
-        return getUserRepo().findAll();
+    @Transactional
+    public void changeUserPassword(String username, String password, boolean asAdmin) {
+        Optional<User> userOptional = getUserRepo().findByUsername(username);
+        if(! userOptional.isPresent()) {
+            throw new UsernameNotFoundException(
+                String.format("No user found with username '%s'.", username));
+        }
+
+        User user = userOptional.get();
+        if (user.getType() == User.VALID_USER_TYPES.REMOTE) {
+            throw new ForbiddenException("Cannot change or set passwords for external users.");
+        }
+
+        //TODO password strength test?
+        user.setPassword(encodePassword(password));
+        getUserRepo().save(user);
     }
 
     @Transactional
@@ -29,7 +53,12 @@ public class LocalUsersDetailsService extends DefaultUsersDetailsService impleme
         user.setPassword(encodePassword(req.getPassword()));
         user.setEnabled(true);
         user.setAdmin(req.isAdmin());
+        user.setType(req.getType());
         return saveUser(user);
+    }
+
+    public String encodePassword(String password) {
+        return (password != null && !password.isEmpty()) ? passwordEncoder.encode(password) : null;
     }
