+++ users/src/main/java/org/openmbee/mms/users/controller/UsersController.java
@@ -8,8 +8,8 @@ import org.openmbee.mms.core.exceptions.UnauthorizedException;
 import org.openmbee.mms.core.utils.AuthenticationUtils;
 import org.openmbee.mms.data.domains.global.User;
 import org.openmbee.mms.users.security.UsersDetailsService;
-import org.openmbee.mms.users.security.UsersCreateRequest;
-import org.openmbee.mms.users.security.UsersResponse;
+import org.openmbee.mms.users.objects.UsersCreateRequest;
+import org.openmbee.mms.users.objects.UsersResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
