+++ users/src/main/java/org/openmbee/mms/users/security/UsersDetailsService.java
@@ -1,6 +1,7 @@
 package org.openmbee.mms.users.security;
 
 import org.openmbee.mms.data.domains.global.User;
+import org.openmbee.mms.users.objects.UsersCreateRequest;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
 import java.util.List;
@@ -16,6 +17,7 @@ public interface UsersDetailsService extends org.springframework.security.core.u
 
     void changeUserPassword(String username, String password, boolean asAdmin);
 
+    String encodePassword(String password);
 
 
     List<User> getUsers();
